Change dialogs and menu bar code structure
Creation of dialog element widgets and their positioning passed to separated classes, they are: DataLayout, ModelLayout and AnalysisLayout. Only the manipulation of information in the layouts remained in the classes DataDialog, ModelDialog and AnalysisDialog;
Dialog Layouts condensed into a new file, dialog_layouts.py;
Created classmethod to get the InsertionDeviceModel object when the model dialog is called.
Created class IMAIDsMenuBar to store the creation of the menu bar and its actions. The signals were kept in the MainWindow.
