+++ gui/widgets/data_dialog.py
@@ -2,22 +2,12 @@
 import os
 import json
 
-from PyQt6.QtWidgets import (QVBoxLayout,
-                             QHBoxLayout,
-                             QGridLayout,
-                             QLineEdit,
-                             QLabel,
-                             QDoubleSpinBox,
-                             QCheckBox,
-                             QDialog,
-                             QDialogButtonBox,
+from PyQt6.QtWidgets import (QDialog,
                              QFileDialog,
-                             QToolButton,
-                             QSpinBox,
                              QMessageBox)
-from PyQt6.QtCore import Qt
 
 from imaids.insertiondevice import InsertionDeviceData
+from .dialog_layouts import DataLayout
 
 class DataDialog(QDialog):
 
@@ -49,56 +39,12 @@ class DataDialog(QDialog):
         self.newfiles = []
         self.lines_names = []
 
-        self.vbox = QVBoxLayout()
+        self.layoutData = DataLayout(parent=self)
 
-        self.grid = QGridLayout()
-
-        label_browse = QLabel("Browse for Data")
-        label_files = QLabel("Files")
-        label_files.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-        label_files.setStyleSheet("font-weight: bold")
-        label_names = QLabel("Names")
-        label_names.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-        label_names.setStyleSheet("font-weight: bold")
-        label_nr_periods = QLabel("Number of Periods")
-        label_nr_periods.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-        label_nr_periods.setStyleSheet("font-weight: bold")
-        label_period_length = QLabel("Period Length")
-        label_period_length.setAlignment(Qt.AlignmentFlag.AlignHCenter)
-        label_period_length.setStyleSheet("font-weight: bold")
-        
-        button_browse = QToolButton()
-        button_browse.setText("...")
-        #button_browse.setSizePolicy(QSizePolicy.Policy.Fixed,QSizePolicy.Policy.Fixed)
-        button_browse.clicked.connect(self.browse)
-
-        self.checkbox_valuesforall = QCheckBox("Use period values change for all")
-        self.checkbox_valuesforall.setChecked(True)
-        
-
-        hbox_browse = QHBoxLayout()
-        hbox_browse.addWidget(label_browse)
-        hbox_browse.addWidget(button_browse)
-
-        self.grid.addLayout(hbox_browse,0,0,Qt.AlignmentFlag.AlignLeft)
-        self.grid.addWidget(self.checkbox_valuesforall,0,2,1,2)
-        self.grid.addWidget(label_files,1,0)
-        self.grid.addWidget(label_names,1,1)
-        self.grid.addWidget(label_nr_periods,1,2)
-        self.grid.addWidget(label_period_length,1,3)
-
-        ## dialog button box - buttons
-        buttons = QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel
-        # dialog button box
-        self.buttonBox = QDialogButtonBox(buttons)
+        self.layoutData.button_browse.clicked.connect(self.browse)
         ## dialog button box - signals
-        self.buttonBox.accepted.connect(self.accept)
-        self.buttonBox.rejected.connect(self.reject)
-        
-        self.vbox.addLayout(self.grid)
-        self.vbox.addWidget(self.buttonBox)
-
-        self.setLayout(self.vbox)
+        self.layoutData.buttonBox.accepted.connect(self.accept)
+        self.layoutData.buttonBox.rejected.connect(self.reject)
 
         #todo: impedir que o dialog seja esticavel para baixo
         #self.setFixedHeight(self.minimumSizeHint().height())
@@ -139,34 +85,27 @@ class DataDialog(QDialog):
                 self.lines_names]
     
     def getUndulatorName(self,filename):
-        Nchars = len(filename)
-        models = self.parameters.keys()
+        models = list(self.parameters.keys())
         
         acertos = []
-        #variar tamanhos de pedacos da string de filename
-        #percorrendo ao contrario para nao pegar trechos pequenos, sem sentido, dos models names
-        for strlen in range(15,4-1,-1):
-            #percorrer o filename e tirando pedacoes de tamanho strlen e procurando se ha models com esse texto
-            for i in range(Nchars-strlen+1):
-                teste = filename[i:i+strlen]
-                #print('teste:',teste)
-                if teste in models:
-                    acertos.append(teste)
-                    break
-                for model in models:
-                    if teste in model:
-                        #print('houve acerto')
-                        acertos.append(teste)
-            else:
-                continue
-            break
-        #print('acertos:',acertos)
-        return acertos[-1]
+        possible_names = ["Delta","Prototype","Sabia"]
+        
+        for modelname in possible_names+models:
+            if modelname in filename:
+                acertos.append(modelname)
+        if acertos:
+            return acertos[-1]
+        else:
+            return ""
     
     def getUndulatorPhase(self, filename):
         phase_idx = filename.find("Phase")
-        phase = filename[phase_idx:].lstrip("Phase")[:2]
-        return phase
+        if phase_idx!=-1:
+            phase = filename[phase_idx:].lstrip("Phase")[:3]
+            phase = "".join([char for char in list(phase) if char.isdigit()])
+            return phase
+        else:
+            return "N"
         
     @classmethod
     def getOpenFileIDs(cls, files=[] ,parent=None):
@@ -208,7 +147,6 @@ class DataDialog(QDialog):
     def browse(self):
 
         userhome = os.path.expanduser('~')
-        
         # filenames: file adress + file name of the selected data files
         files, _ = QFileDialog.getOpenFileNames(parent=self, 
                                                 caption='Open data',
@@ -217,40 +155,32 @@ class DataDialog(QDialog):
         
         loaded_files = self.check_files(files=files)
         
-        
         #todo: criar maneira de poder excluir uma linha depois de importa-la
         nr_oldfiles = len(self.oldfiles)
-        rows = self.grid.rowCount()
-
+        rows = self.layoutData.gridFiles.rowCount()
 
         for i, file in enumerate(loaded_files):
 
             filename = os.path.basename(file)
-            label_file = QLabel(filename)
+            line_name, spin_nr_periods, spin_period_length = self.layoutData.gridFiles_insertAfterRow(filename, i+rows)
 
-            line_name = QLineEdit()
             line_name.textChanged.connect(self.resize_to_content)
-            #todo: mudar name e phase para um padrao caso nao encontre nada no filename
             und_name = self.getUndulatorName(filename)
             und_phase = self.getUndulatorPhase(filename)
-            #line_name.setText(f"Data {(rows-2)+nr_oldfiles+i+1}")
-            line_name.setText(f"{und_name} Phase {und_phase}")
+            if und_name!="":
+                line_name.setText(f"{und_name} Phase {und_phase}")
+            else:
+                line_name.setText(f"Data {(rows-2)+nr_oldfiles+i+1}")
             self.lines_names.append(line_name)
-            
-            spin_nr_periods = QSpinBox()
-            spin_nr_periods.setValue(self.parameters[und_name]["nr_periods"])
+
+            if und_name in self.parameters:
+                spin_nr_periods.setValue(self.parameters[und_name]["nr_periods"])
+                spin_period_length.setValue(self.parameters[und_name]["period_length"])
             spin_nr_periods.valueChanged.connect(self.spin_all)
             self.spins_nr_periods.append(spin_nr_periods)
-            spin_period_length = QDoubleSpinBox()
-            spin_period_length.setValue(self.parameters[und_name]["period_length"])
-            spin_period_length.setSuffix(" mm")
+
             spin_period_length.valueChanged.connect(self.spin_all)
             self.spins_period_length.append(spin_period_length)
-
-            self.grid.addWidget(label_file,i+rows,0)
-            self.grid.addWidget(line_name,i+rows,1)
-            self.grid.addWidget(spin_nr_periods,i+rows,2)
-            self.grid.addWidget(spin_period_length,i+rows,3)
     
     def resize_to_content(self):
         lineedit = self.sender()
@@ -259,7 +189,7 @@ class DataDialog(QDialog):
     def spin_all(self, value):
 
         # realizar o spin all somente quando a checkbox estiver marcada
-        if self.checkbox_valuesforall.isChecked():
+        if self.layoutData.checkbox_valuesforall.isChecked():
         
             # QObject which the signal was sent
             spinbox = self.sender()
